fix(core): fix emulator build and setup issues (#249)

+++ core/src/boot.py
@@ -12,9 +12,10 @@ def clear() -> None:
         storage.wipe()
     if config.has_pin() and config.get_pin_rem() == 0:
         storage.wipe()
-    if storage.device.get_wp_cnts() == 0:
-        for _size, _attrs, name in io.fatfs.listdir("1:/res/wallpapers"):
-            io.fatfs.unlink(f"1:/res/wallpapers/{name}")
+    if not utils.EMULATOR:
+        if storage.device.get_wp_cnts() == 0:
+            for _size, _attrs, name in io.fatfs.listdir("1:/res/wallpapers"):
+                io.fatfs.unlink(f"1:/res/wallpapers/{name}")
 
 
 async def bootscreen() -> None:
